requirements.txt
google-api-python-client==2.184.0
obs-websocket-py==1.0
requests==2.32.5
python-dotenv==1.1.1